requirements.txt
flask==3.0.3
tensorflow==2.20.0
numpy==1.26.4
pillow==10.4.0
Werkzeug == 3.0.0
gunicorn== 21.2.0
matplotlib == 3.10.6
pandas == 2.3.2
scikit-learn == 1.7.1
python-multipart == 0.0.9
tensorflow-serving-api == 2.19.1
fastapi == 0.111.0
uvicorn==0.30.1
